APIClient::item_iterator: reset query_parameters after returning all items
The item_iterator method uses a query parameter called `since` to get the next
results from the API resource when the previous response includes a hasMore=True
property.

The problem here is the last API call leaves the query parameter defined to a
value related to the previous resource retrieved, and if you use the same api
client to iterate over different resources (i.e. campaigns), the results will
always be affected by the previous call.